requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
openai>=1.0
